requirements.txt
alembic==1.6.5
blinker==1.4
certifi==2020.4.5.1
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
flask-heroku==0.1.9
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==19.7.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
markdown2==2.4.0
MarkupSafe==2.0.1
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
urllib3==1.26.5
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.3.3
zipp==3.4.1
